+++ NEWS.md
@@ -1,3 +1,10 @@
+# mgsub 1.5
+
+* Removed NSE evalutaion, deprecating the named list (dictionary) method
+* Added support for vector inputs
+* Added NA handling
+* Slightly improved vectorized performance
+
 # mgsub 1.0.1
 
 * Added checks for named input to mgsub